Refactoring into packages, some string refactoring
