+++ app/src/main/java/infrmr/newsapp/github/com/ifrmr/NavigationDrawerFragment.java
@@ -88,9 +88,6 @@ public class NavigationDrawerFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-
-
-
         if (mDrawerListView != null) {
             if (container != null)
                 container.removeView(mDrawerListView);
@@ -107,11 +104,7 @@ public class NavigationDrawerFragment extends Fragment {
                     getActionBar().getThemedContext(),
                     android.R.layout.simple_list_item_activated_1,
                     android.R.id.text1,
-                    new String[]{
-                            getString(R.string.title_section1),
-                            getString(R.string.title_section2),
-                            getString(R.string.title_section3),
-                    }));
+                    getResources().getStringArray(R.array.title_sections)));
             mDrawerListView.setItemChecked(mCurrentSelectedPosition, true);
 
         } catch (InflateException e) {
@@ -276,7 +269,6 @@ public class NavigationDrawerFragment extends Fragment {
     private void showGlobalContextActionBar() {
         ActionBar actionBar = getActionBar();
         actionBar.setDisplayShowTitleEnabled(true);
-        actionBar.setNavigationMode(ActionBar.NAVIGATION_MODE_STANDARD);
         actionBar.setTitle(R.string.app_name);
     }
