+++ app/src/main/res/values/strings.xml
@@ -14,14 +14,17 @@
     <string name="settings_title">Settings</string>
     <string name="title_activity_article">Article</string>
     <!-- new -->
-    <string name="title_section1">Section 1</string>
-    <string name="title_section2">Section 2</string>
-    <string name="title_section3">Section 3</string>
+    <string-array name="title_sections">
+        <item>Section 1</item>
+        <item>Section 2</item>
+        <item>Section 3</item>
+    </string-array>
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
     <string name="action_example">Example action</string>
-    <string name="loading_news">Loading News...</string>
+    <string name="loading_news">Loading News…</string>
+    <string name="loading">Loading…</string>
 
 </resources>
